Renamed typelookupservice directory. Fix include order.
