+++ src/cpp/CMakeLists.txt
@@ -117,11 +117,11 @@ set(${PROJECT_NAME}_source_files
     fastdds/domain/DomainParticipant.cpp
     fastdds/domain/qos/DomainParticipantQos.cpp
     fastdds/domain/qos/DomainParticipantFactoryQos.cpp
-    fastdds/builtin/typelookupservice/common/TypeLookupTypes.cpp
-    fastdds/builtin/typelookupservice/common/RPCHeadersImpl.cpp
-    fastdds/builtin/typelookupservice/TypeLookupManager.cpp
-    fastdds/builtin/typelookupservice/TypeLookupRequestListener.cpp
-    fastdds/builtin/typelookupservice/TypeLookupReplyListener.cpp
+    fastdds/builtin/type_lookup_service/common/TypeLookupTypes.cpp
+    fastdds/builtin/type_lookup_service/common/RPCHeadersImpl.cpp
+    fastdds/builtin/type_lookup_service/TypeLookupManager.cpp
+    fastdds/builtin/type_lookup_service/TypeLookupRequestListener.cpp
+    fastdds/builtin/type_lookup_service/TypeLookupReplyListener.cpp
     rtps/transport/ChainingTransport.cpp
     rtps/transport/ChannelResource.cpp
     rtps/transport/PortBasedTransportDescriptor.cpp
