+++ src/cpp/fastdds/builtin/type_lookup_service/common/RPCHeadersImplCdrAux.ipp
