+++ src/cpp/fastdds/builtin/type_lookup_service/common/ReplyHeader.hpp
@@ -21,7 +21,7 @@
 #define REPLYHEADER_HPP
 
 #include <fastrtps/rtps/common/SampleIdentity.h>
-#include <fastdds/builtin/typelookupservice/common/Types.hpp>
+#include <fastdds/builtin/type_lookup_service/common/Types.hpp>
 
 namespace eprosima {
