+++ src/cpp/fastdds/builtin/type_lookup_service/common/TypeLookupTypesCdrAux.ipp
